+++ changelog.md
@@ -1,3 +1,1 @@
-# Moni Labs 0.13.5
-
-* Flip behavior of --relative flag when using the jar as a script to better match expectations
+# c